+++ js/render-fullsize.js
@@ -5,7 +5,7 @@ const bigPictureModal = document.querySelector('.big-picture');
 const bigPictureImg = bigPictureModal.querySelector('.big-picture__img');
 const photoDescription = bigPictureModal.querySelector('.social__caption');
 const likesCount = bigPictureModal.querySelector('.likes-count');
-const loadedCommentsCount = document.querySelector('.social__loaded-comments-count');
+const loadedCommentsCount = bigPictureModal.querySelector('.social__loaded-comments-count');
 const commentsCount = bigPictureModal.querySelector('.comments-count');
 const socialCommentsList = bigPictureModal.querySelector('.social__comments');
 const commentsShowMore = bigPictureModal.querySelector('.comments-loader');
@@ -19,14 +19,6 @@ const onBigPhotoEscKeydown = (evt) => {
   }
 };
 
-function closeBigPictureBlock () {
-  body.classList.remove('modal-open');
-  bigPictureModal.classList.add('hidden');
-
-  bigPictureClose.removeEventListener('click', closeBigPictureBlock);
-  document.removeEventListener('keydown', onBigPhotoEscKeydown);
-}
-
 function openBigPictureBlock () {
   body.classList.add('modal-open');
   bigPictureModal.classList.remove('hidden');
@@ -35,12 +27,20 @@ function openBigPictureBlock () {
   document.addEventListener('keydown', onBigPhotoEscKeydown);
 }
 
+function closeBigPictureBlock () {
+  body.classList.remove('modal-open');
+  bigPictureModal.classList.add('hidden');
+
+  bigPictureClose.removeEventListener('click', closeBigPictureBlock);
+  document.removeEventListener('keydown', onBigPhotoEscKeydown);
+}
+
 const renderBigPictureBlock = (smallPicture) => {
   bigPictureImg.querySelector('img').src = smallPicture.url;
   likesCount.textContent = smallPicture.likes;
   photoDescription.textContent = smallPicture.description;
 
-  const { comments } = smallPicture;
+  const {comments} = smallPicture;
   commentsCount.textContent = comments.length;
   socialCommentsList.innerHTML = '';
